rename deployer to gateway
