+++ protostar/commands/deploy/deploy_command.py
@@ -2,9 +2,11 @@ from logging import Logger
 from typing import List, Optional
 
 from protostar.cli.command import Command
-from protostar.deployer import Deployer
-from protostar.deployer.deployer import InvalidNetworkConfigurationException
-from protostar.deployer.network_config import NetworkConfig
+from protostar.starknet_gateway import GatewayFacade
+from protostar.starknet_gateway.gateway_facade import (
+    InvalidNetworkConfigurationException,
+)
+from protostar.starknet_gateway.network_config import NetworkConfig
 from protostar.protostar_exception import ProtostarException
 
 
@@ -33,8 +35,8 @@ class DeployCommand(Command):
         type="str",
     )
 
-    def __init__(self, deployer: Deployer, logger: Logger) -> None:
-        self._deployer = deployer
+    def __init__(self, gateway_facade: GatewayFacade, logger: Logger) -> None:
+        self._gateway_facade = gateway_facade
         self._logger = logger
 
     @property
@@ -96,7 +98,7 @@ class DeployCommand(Command):
 
     async def run(self, args):
         try:
-            network_config = self._deployer.build_network_config(
+            network_config = self._gateway_facade.build_network_config(
                 network=args.network, gateway_url=args.gateway_url
             )
         except InvalidNetworkConfigurationException as err:
@@ -104,7 +106,7 @@ class DeployCommand(Command):
                 f"Argument `{DeployCommand.gateway_url_arg.name}` or `{DeployCommand.network_arg.name}` is required"
             ) from err
 
-        response = await self._deployer.deploy(
+        response = await self._gateway_facade.deploy(
             compiled_contract_path=args.contract,
             gateway_url=network_config.gateway_url,
             inputs=args.inputs,
