+++ protostar/commands/test/cheatcodes/prepare_cheatcode.py
@@ -7,8 +7,8 @@ from starkware.starknet.core.os.contract_address.contract_address import (
 )
 
 from protostar.commands.test.cheatcodes.declare_cheatcode import DeclaredContract
-from protostar.starknet.cheatcode import Cheatcode
 from protostar.commands.test.test_environment_exceptions import CheatcodeException
+from protostar.starknet.cheatcode import Cheatcode
 from protostar.utils.data_transformer_facade import DataTransformerFacade
