--- protostar/deployer/__init__.py
@@ -1,1 +1,0 @@
-from protostar.deployer.deployer import Deployer
