+++ protostar/starknet_gateway/__init__.py
@@ -1,0 +1,1 @@
+from protostar.starknet_gateway.gateway_facade import GatewayFacade
