+++ tests/integration/deploying/declare_command_test.py
@@ -5,7 +5,7 @@ import pytest
 from pytest_mock import MockerFixture
 
 from protostar.commands.declare.declare_command import DeclareCommand
-from protostar.deployer import Deployer
+from protostar.starknet_gateway import GatewayFacade
 
 
 @pytest.mark.parametrize("contract_name", ["main_with_constructor"])
@@ -16,7 +16,7 @@ async def test_deploying_contract(
     compiled_contract_filepath,
 ):
     declare_command = DeclareCommand(
-        deployer=Deployer(project_root_path), logger=mocker.MagicMock()
+        gateway_facade=GatewayFacade(project_root_path), logger=mocker.MagicMock()
     )
     args = SimpleNamespace()
     args.contract = compiled_contract_filepath
